Load IR bear animation from SD within the infrared app
