--- src/MicroPython/picoware/gui/ir_bear_frames.py
@@ -1,377 +1,0 @@
-"""Reviewed Picoware bears: 48x48, row-major MSB-first, 0=ink, 1=background.
-
-Listening: 12 frames; saved: 8 frames; no signal: 6 frames. Playback: 5 fps.
-"""
-
-LISTENING = (
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xf7\xff\xff\xff\xff\xff\xdf\xf1\xff\xff\xff\xff\xbf\xec\x0c\x7f\xff\xff\xf3\xa8\x01\xbf\xff"
-        b"\xff\x6f\x40\x61\xbf\xff\xff\x7e\x20\x00\x7f\xff\xff\x7c\x0f\x40\x7f\xff\xff\xfc\x17\x80\x7f\xff"
-        b"\xff\xfc\x47\x80\x7f\xff\xff\xf8\x6f\xc0\x3f\xff\xff\xe0\x7f\xc0\x3f\xff\xff\xc0\x5f\xc0\x1f\xff"
-        b"\xff\x80\x4f\x80\x8f\xff\xff\x00\x00\x01\x03\xff\xff\x00\x80\x00\x01\xff\xfe\x00\x00\x00\x00\x7f"
-        b"\xfe\x00\x40\x01\x80\x3f\xfe\x00\x70\x07\x80\x1f\xff\x00\x7c\x3f\x80\x1f\xff\x80\x7e\x7f\x80\x0f"
-        b"\xff\xc0\x3f\xff\x00\x0f\xff\xff\xbf\xfe\x00\x0f\xff\xff\x8f\xf8\x10\x0f\xff\xff\x87\xe0\x00\x0f"
-        b"\xff\xff\x81\xc0\x20\x1f\xff\xff\x81\x80\x00\x3f\xff\xff\x80\x00\x00\xff\xff\xff\xc0\x00\x02\xff"
-        b"\xff\xff\xc0\x00\x03\xff\xff\xff\x80\x00\x20\xff\xff\xff\x90\x00\x10\xff\xff\xff\x00\x00\x0f\xff"
-        b"\xff\xff\x00\x00\x07\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x30\x03\xff\xff\xff\x80\xf8\x03\xff"
-        b"\xff\xff\x80\xfc\x03\xff\xff\xff\x80\xfe\x03\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xe7\xff\xff\xff\xff\xff\xdf\xf1\xff\xff\xff\xff\xbf\xec\x0c\x7f\xff\xff\x7f\xa8\x01\xbf\xff"
-        b"\xff\x7f\x40\x61\xbf\xff\xff\x5e\x20\x00\x7f\xff\xff\x5c\x0f\x40\x7f\xff\xff\x7c\x17\x80\x7f\xff"
-        b"\xff\xfc\x47\x80\x7f\xff\xff\xf8\x6f\xc0\x3f\xff\xff\xe0\x7f\xc0\x3f\xff\xff\xc0\x5f\xc0\x1f\xff"
-        b"\xff\x80\x4f\x80\x8f\xff\xff\x00\x00\x01\x03\xff\xff\x00\x80\x00\x01\xff\xfe\x00\x00\x00\x00\x7f"
-        b"\xfe\x00\x40\x01\x80\x3f\xfe\x00\x70\x07\x80\x1f\xff\x00\x7c\x3f\x80\x1f\xff\x80\x7e\x7f\x80\x0f"
-        b"\xff\xc0\x3f\xff\x00\x0f\xff\xff\xbf\xfe\x00\x0f\xff\xff\x8f\xfc\x10\x0f\xff\xff\x87\xe0\x00\x0f"
-        b"\xff\xff\x81\xc0\x20\x1f\xff\xff\x81\x80\x00\x3f\xff\xff\x80\x00\x00\xff\xff\xff\xc0\x00\x02\xff"
-        b"\xff\xff\xc0\x00\x03\xff\xff\xff\x80\x00\x20\xff\xff\xff\x90\x00\x10\xff\xff\xff\x00\x00\x0f\xff"
-        b"\xff\xff\x00\x00\x07\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x30\x03\xff\xff\xff\x80\xf8\x03\xff"
-        b"\xff\xff\x80\xfc\x03\xff\xff\xff\x80\xfe\x03\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xe7\xff\xff\xff\xff\xff\xff\xf3\xff\xff\xff\xff\xff\xec\x0c\x7f\xff\xff\x6f\xa8\x01\xbf\xff"
-        b"\xff\xff\x40\x61\xbf\xff\xff\xde\x20\x00\x7f\xff\xff\xdc\x0f\x40\x7f\xff\xff\x7c\x17\x80\x7f\xff"
-        b"\xff\xfc\x47\x80\x7f\xff\xff\xf8\x6f\xc0\x3f\xff\xff\xe0\x7f\xc0\x3f\xff\xff\xc0\x5f\xc0\x1f\xff"
-        b"\xff\x80\x4f\x80\x8f\xff\xff\x00\x00\x01\x03\xff\xff\x00\x80\x00\x01\xff\xfe\x00\x00\x00\x00\x7f"
-        b"\xfe\x00\x40\x01\x80\x3f\xfe\x00\x70\x07\x80\x3f\xff\x00\x7c\x3f\x80\x1f\xff\x80\x7e\x7f\x80\x0f"
-        b"\xff\xc0\x3f\xff\x00\x0f\xff\xff\xbf\xfe\x00\x0f\xff\xff\x8f\xfc\x10\x0f\xff\xff\x87\xe0\x00\x0f"
-        b"\xff\xff\x81\xc0\x20\x1f\xff\xff\x81\x80\x00\x3f\xff\xff\x80\x00\x00\xff\xff\xff\xc0\x00\x02\xff"
-        b"\xff\xff\xc0\x00\x03\xff\xff\xff\x80\x00\x20\xff\xff\xff\x90\x00\x10\xff\xff\xff\x00\x00\x0f\xff"
-        b"\xff\xff\x00\x00\x07\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x30\x03\xff\xff\xff\x80\xf8\x03\xff"
-        b"\xff\xff\x80\xfc\x03\xff\xff\xff\x80\xfe\x03\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xe7\xff\xff\xff\xff\xff\xbf\xf3\xff\xff\xff\xff\x73\xec\x0c\x7f\xff\xff\x6f\xa8\x01\xbf\xff"
-        b"\xff\xdf\x40\x61\xbf\xff\xff\xde\x20\x00\x7f\xff\xff\xdc\x0f\x40\x7f\xff\xff\x7c\x17\x80\x7f\xff"
-        b"\xff\xfc\x47\x80\x7f\xff\xff\xf8\x6f\xc0\x3f\xff\xff\xe0\x7f\xc0\x3f\xff\xff\xc0\x5f\xc0\x1f\xff"
-        b"\xff\x80\x4f\x80\x8f\xff\xff\x00\x00\x01\x03\xff\xff\x00\x80\x00\x01\xff\xfe\x00\x00\x00\x00\x7f"
-        b"\xfe\x00\x40\x01\x80\x3f\xfe\x00\x70\x07\x80\x3f\xff\x00\x7c\x3f\x80\x1f\xff\x80\x7e\x7f\x80\x0f"
-        b"\xff\xc0\x3f\xff\x00\x0f\xff\xff\xbf\xfe\x00\x0f\xff\xff\x8f\xfc\x10\x0f\xff\xff\x87\xe0\x00\x0f"
-        b"\xff\xff\x81\xc0\x20\x1f\xff\xff\x81\x80\x00\x3f\xff\xff\x80\x00\x00\xff\xff\xff\xc0\x00\x02\xff"
-        b"\xff\xff\xc0\x00\x03\xff\xff\xff\x80\x00\x20\xff\xff\xff\x90\x00\x10\xff\xff\xff\x00\x00\x0f\xff"
-        b"\xff\xff\x00\x00\x07\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x30\x03\xff\xff\xff\x80\xf8\x03\xff"
-        b"\xff\xff\x80\xfc\x03\xff\xff\xff\x80\xfe\x03\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xdf\xff\xff\xff\xff\xff\xbf\xf1\xff\xff\xff\xff\x73\xec\x0c\x7f\xff\xff\xff\xa8\x01\xbf\xff"
-        b"\xfe\xdf\x40\x61\xbf\xff\xfe\xde\x20\x00\x7f\xff\xfe\xdc\x0f\x40\x7f\xff\xff\xdc\x17\x80\x7f\xff"
-        b"\xff\xfc\x47\x80\x7f\xff\xff\xf8\x6f\xc0\x3f\xff\xff\xe0\x7f\xc0\x3f\xff\xff\xc0\x1f\xc0\x1f\xff"
-        b"\xff\x80\x4f\x80\x8f\xff\xff\x00\x00\x01\x03\xff\xff\x00\x80\x00\x01\xff\xfe\x00\x00\x00\x00\x7f"
-        b"\xfe\x00\x40\x01\x80\x3f\xfe\x00\x70\x07\x80\x3f\xff\x00\x7c\x3f\x80\x1f\xff\x80\x7e\x7f\x80\x0f"
-        b"\xff\xc0\x3f\xff\x00\x0f\xff\xff\xbf\xfe\x00\x0f\xff\xff\x8f\xf8\x10\x0f\xff\xff\x87\xe0\x00\x0f"
-        b"\xff\xff\x81\xc0\x20\x1f\xff\xff\x81\x80\x00\x3f\xff\xff\x80\x00\x00\xff\xff\xff\xc0\x00\x02\xff"
-        b"\xff\xff\xc0\x00\x03\xff\xff\xff\x80\x00\x20\xff\xff\xff\x90\x00\x10\xff\xff\xff\x00\x00\x0f\xff"
-        b"\xff\xff\x00\x00\x07\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x30\x03\xff\xff\xff\x80\xf8\x03\xff"
-        b"\xff\xff\x80\xfc\x03\xff\xff\xff\x80\xfe\x03\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xdf\xff\xff\xff\xff\xff\xbf\xf1\xff\xff\xff\xff\x6f\xec\x0c\x7f\xff\xff\xdf\xa8\x01\xbf\xff"
-        b"\xfe\xdf\x40\x61\xbf\xff\xfe\xfe\x20\x00\x7f\xff\xfe\xfc\x0f\x40\x7f\xff\xff\xdc\x17\x80\x7f\xff"
-        b"\xff\x7c\x47\x80\x7f\xff\xff\xf8\x6f\xc0\x3f\xff\xff\xe0\x7f\xc0\x3f\xff\xff\xc0\x1f\xc0\x1f\xff"
-        b"\xff\x80\x4f\x80\x8f\xff\xff\x00\x00\x01\x03\xff\xff\x00\x80\x00\x01\xff\xfe\x00\x00\x00\x00\x7f"
-        b"\xfe\x00\x40\x01\x80\x3f\xfe\x00\x70\x07\x80\x3f\xff\x00\x7c\x3f\x80\x1f\xff\x80\x7e\x7f\x80\x0f"
-        b"\xff\xc0\x3f\xff\x00\x0f\xff\xff\xbf\xfe\x00\x0f\xff\xff\x8f\xf8\x10\x0f\xff\xff\x87\xe0\x00\x0f"
-        b"\xff\xff\x81\xc0\x20\x1f\xff\xff\x81\x80\x00\x3f\xff\xff\x80\x00\x00\xff\xff\xff\xc0\x00\x02\xff"
-        b"\xff\xff\xc0\x00\x03\xff\xff\xff\x80\x00\x20\xff\xff\xff\x90\x00\x10\xff\xff\xff\x00\x00\x0f\xff"
-        b"\xff\xff\x00\x00\x07\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x30\x03\xff\xff\xff\x80\xf8\x03\xff"
-        b"\xff\xff\x80\xfc\x03\xff\xff\xff\x80\xfe\x03\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xef\xff\xff\xff\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\x7f\xf1\xff\xff\xff\xff\xef\xec\x0c\x7f\xff\xfe\xdf\xa8\x01\xbf\xff"
-        b"\xfe\xff\x40\x61\xbf\xff\xfe\xfe\x20\x00\x7f\xff\xfe\xfc\x0f\x40\x7f\xff\xfe\xdc\x17\x80\x7f\xff"
-        b"\xff\x7c\x47\x80\x7f\xff\xff\xf8\x6f\xc0\x3f\xff\xff\xe0\x7f\xc0\x3f\xff\xff\xc0\x1f\xc0\x1f\xff"
-        b"\xff\x80\x4f\x80\x8f\xff\xff\x00\x00\x01\x03\xff\xff\x00\x80\x00\x01\xff\xfe\x00\x00\x00\x00\x7f"
-        b"\xfe\x00\x40\x01\x80\x3f\xfe\x00\x70\x07\x80\x3f\xff\x00\x7c\x3f\x80\x1f\xff\x80\x7e\x7f\x80\x0f"
-        b"\xff\xc0\x3f\xff\x00\x0f\xff\xff\xbf\xfe\x00\x0f\xff\xff\x8f\xf8\x10\x0f\xff\xff\x87\xe0\x00\x0f"
-        b"\xff\xff\x81\xc0\x20\x1f\xff\xff\x81\x80\x00\x3f\xff\xff\x80\x00\x00\xff\xff\xff\xc0\x00\x02\xff"
-        b"\xff\xff\xc0\x00\x03\xff\xff\xff\x80\x00\x20\xff\xff\xff\x90\x00\x10\xff\xff\xff\x00\x00\x07\xff"
-        b"\xff\xff\x00\x00\x07\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x30\x03\xff\xff\xff\x80\xf8\x03\xff"
-        b"\xff\xff\x80\xfc\x03\xff\xff\xff\x80\xfe\x03\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xef\xff\xff\xff\xff"
-        b"\xff\xbf\xff\xff\xff\xff\xff\x73\xf1\xff\xff\xff\xff\xef\xec\x0c\x7f\xff\xfe\xdf\xa8\x01\xbf\xff"
-        b"\xfe\xff\x40\x61\xbf\xff\xfe\xbe\x20\x00\x7f\xff\xfe\xbc\x0f\x40\x7f\xff\xfe\xdc\x17\x80\x7f\xff"
-        b"\xff\xfc\x47\x80\x7f\xff\xff\xf8\x6f\xc0\x3f\xff\xff\xe0\x7f\xc0\x3f\xff\xff\xc0\x1f\xc0\x1f\xff"
-        b"\xff\x80\x4f\x80\x8f\xff\xff\x00\x00\x01\x03\xff\xff\x00\x80\x00\x01\xff\xfe\x00\x00\x00\x00\x7f"
-        b"\xfe\x00\x40\x01\x80\x3f\xfe\x00\x70\x07\x80\x3f\xff\x00\x7c\x3f\x80\x1f\xff\x80\x7e\x7f\x80\x0f"
-        b"\xff\xc0\x3f\xff\x00\x0f\xff\xff\xbf\xfe\x00\x0f\xff\xff\x8f\xf8\x10\x0f\xff\xff\x87\xe0\x00\x0f"
-        b"\xff\xff\x81\xc0\x20\x1f\xff\xff\x81\x80\x00\x3f\xff\xff\x80\x00\x00\xff\xff\xff\xc0\x00\x02\xff"
-        b"\xff\xff\xc0\x00\x03\xff\xff\xff\x80\x00\x20\xff\xff\xff\x90\x00\x10\xff\xff\xff\x00\x00\x07\xff"
-        b"\xff\xff\x00\x00\x07\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x30\x03\xff\xff\xff\x80\xf8\x03\xff"
-        b"\xff\xff\x80\xfc\x03\xff\xff\xff\x80\xfe\x03\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xdf\xff\xff\xff\xff"
-        b"\xff\x7f\xff\xff\xff\xff\xff\xe3\xf1\xff\xff\xff\xfe\xdf\xec\x0c\x7f\xff\xff\xff\xa8\x01\xbf\xff"
-        b"\xfd\xbf\x40\x61\xbf\xff\xfd\xbe\x20\x00\x7f\xff\xfd\xbc\x0f\x40\x7f\xff\xff\xbc\x17\x80\x7f\xff"
-        b"\xfe\xfc\x47\x80\x7f\xff\xff\xf8\x6f\xc0\x3f\xff\xff\xe0\x7f\xc0\x3f\xff\xff\xc0\x1f\xc0\x1f\xff"
-        b"\xff\x80\x4f\x80\x8f\xff\xff\x00\x00\x01\x03\xff\xff\x00\x80\x00\x01\xff\xfe\x00\x00\x00\x00\xff"
-        b"\xfe\x00\x40\x01\x80\x3f\xfe\x00\x70\x07\x80\x3f\xff\x00\x7c\x3f\x80\x1f\xff\x80\x7e\xff\x80\x0f"
-        b"\xff\xc0\x3f\xff\x00\x0f\xff\xff\xbf\xfe\x08\x0f\xff\xff\x8f\xf8\x10\x0f\xff\xff\x87\xe0\x00\x0f"
-        b"\xff\xff\x81\xc0\x20\x1f\xff\xff\x81\x80\x00\x3f\xff\xff\x80\x00\x00\xff\xff\xff\xc0\x00\x02\xff"
-        b"\xff\xff\xc0\x00\x03\xff\xff\xff\x80\x00\x20\xff\xff\xff\x90\x00\x10\xff\xff\xff\x00\x00\x0f\xff"
-        b"\xff\xff\x00\x00\x07\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x30\x03\xff\xff\xff\x80\xf8\x03\xff"
-        b"\xff\xff\x80\xfc\x03\xff\xff\xff\x80\xfe\x03\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xbf\xff\xff\xff\xff"
-        b"\xff\x7f\xff\xff\xff\xff\xfe\xef\xf1\xff\xff\xff\xff\xdf\xec\x0c\x7f\xff\xfd\xbf\xa8\x01\xbf\xff"
-        b"\xfd\xbf\x40\x61\xbf\xff\xfd\xfe\x20\x00\x7f\xff\xfd\xbc\x0f\x40\x7f\xff\xfd\xbc\x1f\x80\x7f\xff"
-        b"\xfe\xfc\x47\x80\x7f\xff\xff\xf8\x6f\xc0\x3f\xff\xff\xe0\x7f\xc0\x3f\xff\xff\xc0\x1f\xc0\x1f\xff"
-        b"\xff\x80\x4f\x80\x8f\xff\xff\x00\x00\x01\x03\xff\xff\x00\x00\x00\x01\xff\xfe\x00\x00\x00\x00\x7f"
-        b"\xfe\x00\x40\x01\x80\x3f\xfe\x00\x70\x07\x80\x3f\xff\x00\x7c\x3f\x80\x1f\xff\x80\x7e\x7f\x80\x0f"
-        b"\xff\xc0\x3f\xff\x00\x0f\xff\xff\xbf\xfe\x00\x0f\xff\xff\x8f\xfc\x10\x0f\xff\xff\x87\xe0\x00\x0f"
-        b"\xff\xff\x81\xc0\x20\x1f\xff\xff\x81\x80\x00\x3f\xff\xff\x80\x00\x00\xff\xff\xff\xc0\x00\x02\xff"
-        b"\xff\xff\xc0\x00\x03\xff\xff\xff\x80\x00\x20\xff\xff\xff\x90\x00\x10\xff\xff\xff\x00\x00\x0f\xff"
-        b"\xff\xff\x00\x00\x07\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x30\x03\xff\xff\xff\x80\xf8\x03\xff"
-        b"\xff\xff\x80\xfc\x03\xff\xff\xff\x80\xfe\x03\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xbf\xff\xff\xff\xff"
-        b"\xff\x7f\xff\xff\xff\xff\xfe\xef\xf1\xff\xff\xff\xff\xff\xec\x0c\x7f\xff\xfd\xbf\xa8\x01\xbf\xff"
-        b"\xfd\xff\x40\x61\xbf\xff\xfd\xfe\x20\x00\x7f\xff\xfd\xfc\x0f\x40\x7f\xff\xfd\xbc\x1f\x80\x7f\xff"
-        b"\xff\xfc\x47\x80\x7f\xff\xff\xf8\x6f\xc0\x3f\xff\xff\xe0\x7f\xc0\x3f\xff\xff\xc0\x1f\xc0\x1f\xff"
-        b"\xff\x80\x4f\x80\x8f\xff\xff\x00\x20\x01\x03\xff\xff\x00\x00\x00\x01\xff\xfe\x00\x00\x00\x00\x7f"
-        b"\xfe\x00\x40\x01\x80\x3f\xfe\x00\x70\x07\x80\x1f\xff\x00\x7c\x3f\x80\x1f\xff\x80\x7e\x7f\x80\x0f"
-        b"\xff\xc0\x3f\xff\x00\x0f\xff\xff\xbf\xfe\x00\x0f\xff\xff\x8f\xfc\x10\x0f\xff\xff\x87\xe0\x00\x0f"
-        b"\xff\xff\x81\xc0\x20\x1f\xff\xff\x81\x80\x00\x3f\xff\xff\x80\x00\x00\xff\xff\xff\xc0\x00\x02\xff"
-        b"\xff\xff\xc0\x00\x03\xff\xff\xff\x80\x00\x20\xff\xff\xff\x90\x00\x10\xff\xff\xff\x00\x00\x0f\xff"
-        b"\xff\xff\x00\x00\x07\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x30\x03\xff\xff\xff\x80\xf8\x03\xff"
-        b"\xff\xff\x80\xfc\x03\xff\xff\xff\x80\xfe\x03\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xdf\xff\xff\xff\xff\xff\xbf\xff\xff\xff\xff"
-        b"\xff\xff\xff\xff\xff\xff\xfe\xff\xf1\xff\xff\xff\xfd\xff\xec\x0c\x7f\xff\xfd\xbf\xa8\x01\xbf\xff"
-        b"\xff\xff\x40\x61\xbf\xff\xff\xfe\x20\x00\x7f\xff\xff\xfc\x0f\x40\x7f\xff\xfd\xac\x1f\x80\x7f\xff"
-        b"\xfd\xfc\x47\x80\x7f\xff\xff\xf8\x6f\xc0\x3f\xff\xff\xe0\x7f\xc0\x3f\xff\xff\xc0\x1f\xc0\x1f\xff"
-        b"\xff\x80\x4f\x80\x8f\xff\xff\x00\x00\x01\x03\xff\xff\x00\x00\x00\x01\xff\xfe\x00\x00\x00\x00\x7f"
-        b"\xfe\x00\x40\x01\x80\x3f\xfe\x00\x70\x07\x80\x1f\xff\x00\x7c\x3f\x80\x1f\xff\x80\x7e\x7f\x80\x0f"
-        b"\xff\xc0\x3f\xff\x00\x0f\xff\xff\xbf\xff\x10\x0f\xff\xff\x8f\xf8\x10\x0f\xff\xff\x87\xe0\x00\x0f"
-        b"\xff\xff\x81\xc0\x20\x1f\xff\xff\x81\x80\x00\x3f\xff\xff\x80\x00\x00\xff\xff\xff\xc0\x00\x02\xff"
-        b"\xff\xff\xc0\x00\x02\xff\xff\xff\x80\x00\x20\xff\xff\xff\x90\x00\x10\xff\xff\xff\x00\x00\x0f\xff"
-        b"\xff\xff\x00\x00\x07\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x30\x03\xff\xff\xff\x80\xf8\x03\xff"
-        b"\xff\xff\x80\xfc\x03\xff\xff\xff\x80\xfe\x03\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-)
-
-SAVED = (
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xf7\xff\xff\xff\xff\xff\xff\xf3\xff\xff\xff\xff\xff\xec\x0c\x7f\xff\xff\xff\xa8\x01\xbf\xff"
-        b"\xff\xee\x40\x61\xbf\xff\xff\x7e\x20\x00\x7f\xff\xff\x7c\x07\x40\x7f\xff\xff\xfc\x17\x80\x7f\xff"
-        b"\xff\xfc\x47\x80\x7f\xff\xff\xf8\x6f\xc0\x3f\xff\xff\xe0\x3f\xc0\x3f\xff\xff\xc0\x1f\xc0\x1f\xff"
-        b"\xff\x80\x0f\x80\x87\xff\xff\x00\x00\x01\x03\xff\xff\x00\x00\x00\x01\xff\xfe\x00\x00\x00\x00\xff"
-        b"\xfe\x00\x40\x01\x80\x3f\xfe\x00\x70\x07\x80\x1f\xff\x00\x7c\x3f\x80\x1f\xff\x80\x7e\x7f\x80\x0f"
-        b"\xff\xc0\x3f\xff\x00\x0f\xff\xff\xbf\xfe\x10\x0f\xff\xff\x8f\xfc\x10\x0f\xff\xff\x87\xe0\x00\x0f"
-        b"\xff\xff\x81\xc0\x20\x1f\xff\xff\x80\x80\x00\x3f\xff\xff\x80\x00\x00\xff\xff\xff\xc0\x00\x02\xff"
-        b"\xff\xff\xc0\x00\x22\xff\xff\xff\x80\x00\x20\xff\xff\xff\x80\x00\x10\xff\xff\xff\x80\x00\x0f\xff"
-        b"\xff\xff\x00\x00\x07\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x30\x03\xff\xff\xff\x80\xf8\x03\xff"
-        b"\xff\xff\x80\xfc\x03\xff\xff\xff\x80\xfe\x03\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xe3\xfc\xff\xff\xff\xff\xe8\x01\x7f\xff"
-        b"\xff\xff\xe8\x03\xbf\xff\xff\xfc\x60\x21\x3f\xff\xff\xf8\x00\x00\x7f\xff\xff\xf0\x03\x00\x7f\xff"
-        b"\xff\xf0\x0f\x80\x7f\xff\xff\xe0\x83\xc0\x3f\xff\xff\xc0\x87\xe0\x3f\xff\xff\xc0\xbf\xe0\x1f\xff"
-        b"\xff\x80\x9f\xc0\x0f\xff\xff\x00\x0f\xc0\x03\xff\xff\x00\x02\x00\x01\xff\xfe\x00\x00\x00\x00\xff"
-        b"\xfe\x00\x40\x00\xc0\x3f\xfe\x00\x60\x01\xc0\x1f\xfe\x00\x78\x07\xc0\x1f\xfe\x00\x7e\x3f\x80\x0f"
-        b"\xff\x00\x7f\x7f\x80\x0f\xff\xc0\x3e\xff\x00\x07\xff\xff\x9f\xfe\x00\x0f\xff\xff\x8f\xf8\x10\x0f"
-        b"\xff\xff\x83\xe0\x00\x0f\xff\xff\x81\xc0\x20\x1f\xff\xff\xc0\x00\x20\x7f\xff\xff\xc0\x00\x20\xff"
-        b"\xff\xff\xc0\x00\x23\x7f\xff\xff\x80\x00\x20\x7f\xff\xff\x80\x00\x10\xff\xff\xff\x00\x00\x09\xff"
-        b"\xff\xff\x00\x00\x03\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x78\x03\xff"
-        b"\xff\xff\x80\xfe\x03\xff\xff\xff\x80\xfe\x01\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xe3\x08\x7f\xff"
-        b"\xff\xff\xd8\x03\xbf\xff\xff\xff\xc0\x01\x3f\xff\xff\xff\xe0\x00\x7f\xff\xff\xe3\xe0\x18\x7f\xff"
-        b"\xff\x81\xe0\x80\x7f\xff\xff\x01\xc3\xe0\x3f\xff\xff\x01\xc7\x70\x3f\xff\xff\x01\xc4\x70\x1f\xff"
-        b"\xfe\x07\x86\xf0\x0f\xff\xfe\x3b\x47\xf0\x47\xff\xfe\x36\x23\xe0\x01\xff\xfe\x00\x01\xc0\x00\xff"
-        b"\xfe\x02\x00\x00\x00\x3f\xfe\x00\x60\x00\xc0\x1f\xfe\x00\x70\x01\xc0\x0f\xfe\x00\x7c\x0f\xc0\x0f"
-        b"\xff\x00\x7f\x3f\x80\x07\xff\x00\x3f\x7f\x80\x07\xff\x80\x1f\xff\x08\x07\xff\xe3\x8f\xfc\x00\x07"
-        b"\xff\xff\x83\xf0\x10\x0f\xff\xff\x81\xe0\x20\x1f\xff\xff\xc0\x80\x20\x3f\xff\xff\xc0\x00\x20\xff"
-        b"\xff\xff\xc0\x00\x21\xff\xff\xff\xc0\x00\x10\x7f\xff\xff\x80\x00\x10\x7f\xff\xff\x00\x00\x08\xff"
-        b"\xff\xff\x00\x00\x03\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x00\x01\xff\xff\xff\x00\x78\x01\xff"
-        b"\xff\xff\x00\xfe\x03\xff\xff\xff\x80\xff\x01\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xe3\x18\x7f\xff"
-        b"\xff\xff\xd8\x03\xbf\xff\xff\xff\xc8\x01\x3f\xff\xff\xff\xe0\x00\x7f\xff\xff\xe3\xe0\x18\x7f\xff"
-        b"\xfc\x83\xe0\x80\x7f\xff\xff\x01\xc3\xe0\x3f\xff\xff\x01\xc7\x70\x3f\xff\xfe\x01\xc4\x70\x1f\xff"
-        b"\xfe\x03\x86\xf0\x1f\xff\xfe\x3b\x07\xf0\x47\xff\xfe\x36\x23\xe0\x01\xff\xfe\x00\x01\xc0\x00\xff"
-        b"\xfe\x02\x00\x00\x00\x3f\xfe\x00\x60\x00\xc0\x1f\xfe\x00\x70\x01\xc0\x0f\xfe\x00\x7e\x0f\xc0\x0f"
-        b"\xff\x00\x7f\x3f\x80\x07\xff\x00\x3f\x7f\x80\x07\xff\x80\x1f\xff\x08\x07\xff\xe3\x8f\xfc\x00\x07"
-        b"\xff\xff\x83\xf0\x10\x0f\xff\xff\x81\xe0\x20\x1f\xff\xff\xc0\x80\x20\x3f\xff\xff\xc0\x00\x20\xff"
-        b"\xff\xff\xc0\x00\x21\xff\xff\xff\xc0\x00\x10\x7f\xff\xff\x80\x00\x10\x7f\xff\xff\x00\x00\x08\xff"
-        b"\xff\xff\x00\x00\x03\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x00\x01\xff\xff\xff\x00\x78\x01\xff"
-        b"\xff\xff\x00\xfe\x03\xff\xff\xff\x80\xff\x01\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xe1\xfc\x3f\xff\xff\xff\xec\x01\xbf\xff"
-        b"\xff\xff\xe8\x00\xbf\xff\xff\xff\xe3\x04\x3f\xff\xff\xff\xf0\x00\x7f\xff\xff\xff\xe0\x70\x3f\xff"
-        b"\xff\xdf\xe0\xf8\x3f\xff\xff\x8f\xe1\xb8\x1f\xff\xff\x8f\xc1\x9c\x1f\xff\xfe\x8f\xc1\xfc\x1f\xff"
-        b"\xfc\x0f\x81\xfc\x07\xff\xfc\x2f\x01\xf8\x01\xff\xf8\xae\x00\x60\x00\xff\xf8\x24\x00\x00\x00\x7f"
-        b"\xf8\xe2\x20\x00\x20\x3f\xf8\x40\x30\x00\xe0\x1f\xfc\x40\x3e\x07\xe0\x0f\xfe\x00\x3f\x9f\xc0\x0f"
-        b"\xff\x00\x3f\xff\xc0\x07\xff\x80\x1f\xff\x80\x07\xff\xc0\x0f\x7f\x08\x07\xff\xe0\xc7\xfc\x00\x07"
-        b"\xff\xff\xc1\xf0\x00\x0f\xff\xff\xc0\x60\x20\x1f\xff\xff\xc0\x00\x20\x3f\xff\xff\xc0\x00\x21\xff"
-        b"\xff\xff\xc0\x00\x01\x7f\xff\xff\xd0\x00\x10\x7f\xff\xff\x80\x00\x00\x7f\xff\xff\x80\x00\x07\xff"
-        b"\xff\xff\x80\x00\x03\xff\xff\xff\x00\x00\x03\xff\xff\xff\x00\x00\x03\xff\xff\xff\x80\x7c\x03\xff"
-        b"\xff\xff\x80\xfe\x03\xff\xff\xff\x80\xff\x01\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xf1\xfe\x7f\xff\xff\xff\xe4\x00\xbf\xff\xff\xff\xec\x00\xbf\xff"
-        b"\xff\xff\xe1\x02\x3f\xff\xff\xff\xf0\x00\x3f\xff\xff\xff\xf1\x22\x3f\xff\xff\xff\xe0\x78\x3f\xff"
-        b"\xff\xff\xe0\xec\x1f\xff\xff\x9f\xe1\x8c\x1f\xff\xff\x1f\xc1\xdc\x1f\xff\xff\x9f\xc1\xfc\x0f\xff"
-        b"\xff\x9f\x80\xfc\x27\xff\xf8\x5f\x00\x78\x01\xff\xf8\x6e\x00\x00\x00\xff\xf8\xe4\x00\x00\x00\x7f"
-        b"\xf8\x62\x30\x00\x60\x3f\xf9\xe0\x38\x00\xe0\x1f\xf8\x40\x3f\x07\xe0\x0f\xfc\x40\x3f\x9f\xc0\x07"
-        b"\xff\x00\x1f\xff\xc0\x07\xff\x00\x1f\xff\x80\x07\xff\x80\x07\xff\x08\x07\xff\xe0\xc3\xf8\x10\x07"
-        b"\xff\xff\xc0\xf0\x00\x0f\xff\xff\xc0\x60\x00\x1f\xff\xff\xc0\x00\x00\x3f\xff\xff\xe0\x00\x01\xff"
-        b"\xff\xff\xc0\x00\x11\x7f\xff\xff\xc0\x00\x10\x7f\xff\xff\xc8\x00\x08\xff\xff\xff\x80\x00\x07\xff"
-        b"\xff\xff\x80\x00\x03\xff\xff\xff\x80\x00\x03\xff\xff\xff\x80\x38\x03\xff\xff\xff\x80\x7c\x03\xff"
-        b"\xff\xff\x80\xfe\x03\xff\xff\xff\x80\xff\x01\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xf1\xfe\x3f\xff\xff\xff\xe4\x00\xbf\xff\xff\xff\xec\x00\xbf\xff"
-        b"\xff\xff\xe1\x02\x3f\xff\xff\xff\xf0\x00\x3f\xff\xff\xff\xf1\x32\x3f\xff\xff\xff\xe0\x78\x3f\xff"
-        b"\xff\xff\xe0\xfc\x1f\xff\xff\x9f\xe1\x8c\x1f\xff\xff\x1f\xc1\xdc\x1f\xff\xff\x9f\xc1\xfc\x0f\xff"
-        b"\xff\x9f\x80\xfc\x27\xff\xf8\x5f\x00\x78\x01\xff\xf8\x6e\x00\x00\x00\xff\xf8\xe4\x00\x00\x00\x7f"
-        b"\xf8\x62\x30\x00\x60\x1f\xf9\xe0\x38\x00\xe0\x1f\xf8\x40\x3f\x07\xe0\x0f\xfc\x40\x3f\x9f\xc0\x07"
-        b"\xff\x00\x1f\xff\xc0\x07\xff\x00\x1f\xff\x88\x07\xff\x80\x07\xff\x08\x07\xff\xe0\xc3\xf8\x10\x07"
-        b"\xff\xff\xc0\xf0\x00\x0f\xff\xff\xc0\x60\x00\x1f\xff\xff\xc0\x00\x00\x3f\xff\xff\xe0\x00\x01\xff"
-        b"\xff\xff\xc0\x00\x11\x7f\xff\xff\xc0\x00\x10\x7f\xff\xff\xc8\x00\x08\xff\xff\xff\x80\x00\x07\xff"
-        b"\xff\xff\x80\x00\x03\xff\xff\xff\x80\x00\x03\xff\xff\xff\x80\x38\x03\xff\xff\xff\x80\x7c\x03\xff"
-        b"\xff\xff\x80\xfe\x03\xff\xff\xff\x80\xff\x01\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xf1\xfe\x7f\xff\xff\xff\xe4\x00\xbf\xff\xff\xff\xec\x01\xbf\xff"
-        b"\xff\xff\xe1\x02\x3f\xff\xff\xff\xf0\x00\x3f\xff\xff\xff\xf1\x22\x3f\xff\xff\xff\xe0\x78\x3f\xff"
-        b"\xff\xff\xe0\xfc\x1f\xff\xff\x9f\xe1\x8c\x1f\xff\xff\x1f\xc1\xdc\x1f\xff\xff\x9f\xc1\xfc\x0f\xff"
-        b"\xff\x9f\x80\xfc\x27\xff\xf8\x5f\x00\x78\x01\xff\xf8\x6e\x00\x00\x00\xff\xf8\xe4\x00\x00\x00\x7f"
-        b"\xf8\x62\x30\x00\x60\x1f\xf9\xe0\x38\x00\xe0\x1f\xf8\x40\x3f\x07\xe0\x0f\xfc\xc0\x3f\x9f\xc0\x07"
-        b"\xff\x00\x1f\xff\xc0\x07\xff\x80\x1f\xff\x88\x07\xff\xc0\x07\xff\x08\x07\xff\xe0\xc3\xf8\x00\x07"
-        b"\xff\xff\xc0\xf0\x00\x0f\xff\xff\xc0\x60\x00\x1f\xff\xff\xc0\x00\x00\x3f\xff\xff\xe0\x00\x01\xff"
-        b"\xff\xff\xc0\x00\x11\x7f\xff\xff\xc0\x00\x10\xff\xff\xff\xc8\x00\x0f\xff\xff\xff\x84\x00\x07\xff"
-        b"\xff\xff\x80\x00\x03\xff\xff\xff\x80\x00\x03\xff\xff\xff\x80\x38\x03\xff\xff\xff\x80\x7c\x03\xff"
-        b"\xff\xff\x80\xfe\x03\xff\xff\xff\x80\xff\x01\xff\xff\xff\x00\xff\x00\xff\xff\xfe\x60\xff\x09\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-)
-
-NO_SIGNAL = (
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\x8f\xff\xcf\xff\xff\xff\xaf\xff\x87\xff\xff\xff\xef\xfc\x37\xff\xff\xff\xdf\xe0\x17\xff\xff"
-        b"\xff\xfc\x00\x07\xff\xff\xff\xd9\x00\x07\xff\xff\xff\xfb\x20\x43\xff\xff\xff\xf0\x00\x03\xff\xff"
-        b"\xff\xc0\x27\x81\xff\xff\xff\xc0\x0e\xc1\xff\xff\xff\x84\x08\xe1\xff\xff\xff\x82\x0e\xe0\xff\xff"
-        b"\xff\x06\x0e\xc0\x7f\xff\xff\x04\x0f\x80\x1f\xff\xfe\x04\x00\x00\x0f\xff\xfe\x00\x00\x00\x07\xff"
-        b"\xfc\x08\x00\x01\x03\xff\xfc\x00\x80\x07\x01\xff\xfe\x00\xe0\x1e\x00\xff\xfe\x00\xfc\x7e\x00\x7f"
-        b"\xff\x00\xfe\xfe\x00\x7f\xff\x80\x7d\x7c\x40\x7f\xff\xf0\x3c\x70\x00\x7f\xff\xfe\x0c\xe0\x80\xff"
-        b"\xff\xfe\x07\xc0\x00\xff\xff\xfe\x01\x00\x03\xff\xff\xfe\x01\x00\x07\xff\xff\xfe\x00\x00\x03\xff"
-        b"\xff\xfe\x00\x00\x83\xff\xff\xfe\x00\x00\x47\xff\xff\xfe\x00\x00\x7f\xff\xff\xfc\x00\x00\x3f\xff"
-        b"\xff\xfc\x00\x00\x3f\xff\xff\xf8\x00\x00\x1f\xff\xff\xf8\x00\x00\x1f\xff\xff\xfc\x03\xc0\x3f\xff"
-        b"\xff\xfc\x07\xe0\x3f\xff\xff\xfe\x07\xe0\x7f\xff\xff\xf0\x07\xe0\x0f\xff\xff\xe0\x07\xe0\x07\xff"
-        b"\xff\xed\x0f\xf0\x97\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\x8f\xff\xff\xff\xff\xff\xa7\xff\xc3\xff\xff\xff\xcf\xff\xcb\xff\xff\xff\xdf\x18\x19\xff\xff"
-        b"\xff\xfe\x40\x03\xff\xff\xff\xfe\x40\x03\xff\xff\xff\xfe\x10\x01\xff\xff\xff\xfd\x00\x01\xff\xff"
-        b"\xff\xe1\x03\xc0\xff\xff\xff\xc1\x03\xe0\xff\xff\xff\xcc\x06\x30\xff\xff\xff\x80\x06\x70\xff\xff"
-        b"\xff\x02\x07\x70\x7f\xff\xff\x06\x07\xe0\x1f\xff\xfe\x00\x01\x00\x0f\xff\xfe\x00\x00\x00\x07\xff"
-        b"\xfe\x00\x00\x00\x03\xff\xfe\x00\x80\x03\x01\xff\xfe\x00\xe0\x0f\x00\xff\xfe\x00\xfc\x7f\x00\x7f"
-        b"\xff\x00\x7e\xfe\x00\x7f\xff\xc0\x7e\x7e\x20\x7f\xff\xfc\x1c\x78\x40\x7f\xff\xfe\x0e\x70\x40\x7f"
-        b"\xff\xfe\x07\xc0\x80\xff\xff\xfe\x01\x80\x81\xff\xff\xff\x00\x00\x83\xff\xff\xff\x00\x00\x83\xff"
-        b"\xff\xff\x00\x00\xc3\xff\xff\xfe\x00\x00\x67\xff\xff\xfe\x00\x00\x7f\xff\xff\xfc\x00\x00\x3f\xff"
-        b"\xff\xfc\x00\x00\x3f\xff\xff\xf8\x00\x00\x1f\xff\xff\xf8\x00\x00\x1f\xff\xff\xfc\x03\xc0\x3f\xff"
-        b"\xff\xfc\x07\xe0\x3f\xff\xff\xff\x07\xe0\x7f\xff\xff\xf0\x07\xe0\x0f\xff\xff\xe0\x07\xe0\x07\xff"
-        b"\xff\xe9\x0f\xf0\x97\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xe3\xff\xff\xff\xff\xff\xc9\xff\xff\xff\xff\x10\x09\xff\xff"
-        b"\xff\xfe\x00\x03\xff\xff\xff\xfe\xc0\x03\xff\xff\xff\xfe\x10\x01\xff\xff\xff\xff\x00\x01\xff\xff"
-        b"\xff\xff\x03\xc0\xff\xff\xff\xff\x07\x60\xff\xff\xff\xff\x04\x60\xff\xff\xff\xfe\x06\x70\xff\xff"
-        b"\xff\xfe\x07\x60\x7f\xff\xfe\xfc\x07\xe0\x3f\xff\xe4\x78\x01\x00\x3f\xff\xc1\x38\x00\x00\x1f\xff"
-        b"\xe3\x90\x00\x01\x1f\xff\xe3\x80\xc0\x03\x0f\xff\xf0\x00\xf0\x0f\x0f\xff\xfc\x00\xfc\x7e\x0f\xff"
-        b"\xfc\x00\x7e\xfe\x0c\x7f\xfe\x00\x7d\x7c\x10\xff\xff\x00\x1c\x78\x23\x0f\xff\x82\x0e\x70\x07\x0f"
-        b"\xff\xfe\x07\xc0\x03\x1f\xff\xff\x01\x80\xc0\x3f\xff\xff\x00\x00\xe0\xff\xff\xff\x00\x00\xff\xff"
-        b"\xff\xff\x00\x00\xff\xff\xff\xfe\x00\x00\x7f\xff\xff\xfe\x00\x00\x7f\xff\xff\xfc\x00\x00\x3f\xff"
-        b"\xff\xfc\x00\x00\x3f\xff\xff\xf8\x00\x00\x1f\xff\xff\xf8\x00\x00\x1f\xff\xff\xfc\x03\xc0\x3f\xff"
-        b"\xff\xfc\x07\xe0\x3f\xff\xff\xfe\x07\xe0\x7f\xff\xff\xf0\x07\xe0\x0f\xff\xff\xe0\x07\xe0\x07\xff"
-        b"\xff\xed\x0f\xf0\x97\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xe1\xff\xff\xff\xff\xbc\x0d\xff\xff"
-        b"\xff\xfe\x00\x0d\xff\xff\xff\xfe\xc0\x01\xff\xff\xff\xfe\x48\x01\xff\xff\xff\xff\x00\x01\xff\xff"
-        b"\xff\xff\x08\x90\xff\xff\xff\xff\x0b\xc0\xff\xff\xff\xff\x03\x20\x7f\xff\xff\xfe\x06\x30\x7f\xff"
-        b"\xff\xfe\x07\x70\x7f\xff\xfe\xfc\x07\x70\x3f\x7f\xfe\x38\x03\xe0\x1c\x7f\xe1\x90\x00\x00\x19\x87"
-        b"\xe1\xd0\x00\x00\x0b\x87\xf1\x80\xc0\x03\x01\x8f\xf8\x00\xf0\x0f\x00\x1f\xfe\x00\x7c\x3f\x00\x7f"
-        b"\xfe\x00\x7f\xfe\x00\xff\xff\x00\x3e\x7e\x00\xff\xff\x80\x1c\x38\x01\xff\xff\xc7\x0e\x70\xe3\xff"
-        b"\xff\xff\x03\xc0\xff\xff\xff\xff\x01\x80\xff\xff\xff\xff\x00\x00\xff\xff\xff\xff\x00\x00\xff\xff"
-        b"\xff\xff\x00\x00\xff\xff\xff\xfe\x00\x00\x7f\xff\xff\xfe\x00\x00\x7f\xff\xff\xfc\x00\x00\x3f\xff"
-        b"\xff\xfc\x00\x00\x3f\xff\xff\xfc\x00\x00\x1f\xff\xff\xf8\x00\x00\x1f\xff\xff\xf8\x03\xc0\x1f\xff"
-        b"\xff\xfc\x07\xe0\x3f\xff\xff\xfe\x07\xe0\x7f\xff\xff\xf0\x07\xe0\x0f\xff\xff\xe0\x07\xe0\x07\xff"
-        b"\xff\xe5\x0f\xf0\x97\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xe1\xff\xff\xff\xff\xbc\x0d\xff\xff"
-        b"\xff\xfe\x00\x0d\xff\xff\xff\xfe\xc0\x21\xff\xff\xff\xfe\x48\x01\xff\xff\xff\xff\x00\x01\xff\xff"
-        b"\xff\xff\x08\x00\xff\xff\xff\xff\x0b\xc0\xff\xff\xff\xff\x06\x60\xff\xff\xff\xfe\x06\x70\x7f\xff"
-        b"\xfe\xfe\x07\xf0\x7f\x7f\xfe\x7c\x06\xe0\x3e\x7f\xe1\x38\x07\xc0\x1c\x87\xe1\x90\x00\x00\x09\x97"
-        b"\xe1\x80\x00\x00\x01\x87\xf0\x08\xc0\x03\x00\x0f\xfc\x00\xf0\x0f\x00\x3f\xfe\x00\xfe\x7f\x00\x7f"
-        b"\xff\x00\x7f\xfe\x00\xff\xff\x00\x3c\x7c\x00\xff\xff\x80\x1d\x78\x01\xff\xff\xc3\x0e\x70\x43\xff"
-        b"\xff\xff\x03\xc0\x7f\xff\xff\xff\x01\x80\xff\xff\xff\xff\x00\x00\xff\xff\xff\xff\x00\x00\xff\xff"
-        b"\xff\xff\x00\x00\xff\xff\xff\xfe\x00\x00\x7f\xff\xff\xfe\x00\x00\x7f\xff\xff\xfc\x00\x00\x3f\xff"
-        b"\xff\xfc\x00\x00\x3f\xff\xff\xf8\x00\x00\x1f\xff\xff\xf8\x00\x00\x1f\xff\xff\xf8\x03\xc0\x1f\xff"
-        b"\xff\xfc\x07\xe0\x3f\xff\xff\xfe\x07\xe0\x3f\xff\xff\xf0\x0f\xe0\x0f\xff\xff\xe0\x0f\xe0\x07\xff"
-        b"\xff\xe9\x0f\xf0\x97\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-    (
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-        b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xf3\xff\xff\xff\xff\xbf\x21\xff\xff"
-        b"\xff\xff\x00\x0d\xff\xff\xff\xfe\x40\x01\xff\xff\xff\xfe\x48\x01\xff\xff\xff\xff\x00\x01\xff\xff"
-        b"\xff\xff\x08\x10\xff\xff\xff\xff\x0b\xc0\xff\xff\xff\xff\x07\xe0\xff\xff\xff\xfe\x06\x30\x7f\xff"
-        b"\xfd\xfe\x06\x70\x7f\x7f\xf8\xf8\x07\x70\x3e\x3f\xc6\x70\x07\xe0\x1c\xe7\xd3\x30\x00\x00\x09\x97"
-        b"\xc7\x00\x00\x00\x09\xc7\xe0\x00\x80\x01\x00\x0f\xf0\x00\xe0\x0f\x00\x1f\xfc\x00\xfc\x3f\x00\x7f"
-        b"\xfc\x00\x7e\xfe\x00\x7f\xfe\x00\x7e\x7e\x00\xff\xff\x00\x3c\x7c\x00\xff\xff\x82\x0c\x70\x01\xff"
-        b"\xff\xce\x07\xc0\x67\xff\xff\xfe\x01\x80\xff\xff\xff\xff\x01\x00\xff\xff\xff\xff\x00\x00\xff\xff"
-        b"\xff\xfe\x00\x00\xff\xff\xff\xfe\x00\x00\x7f\xff\xff\xfe\x00\x00\x7f\xff\xff\xfc\x00\x00\x3f\xff"
-        b"\xff\xfc\x00\x00\x3f\xff\xff\xf8\x00\x00\x3f\xff\xff\xf8\x00\x00\x1f\xff\xff\xf8\x03\xc0\x1f\xff"
-        b"\xff\xfc\x07\xe0\x3f\xff\xff\xfc\x07\xe0\x7f\xff\xff\xf0\x07\xe0\x0f\xff\xff\xe0\x07\xe0\x07\xff"
-        b"\xff\xed\x0f\xf0\xb7\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff\xff"
-    ),
-)
